Discarding No-Autofocus Changes
Attempted lots of fixes, but no good solution has been found for
the time being. Ideally, the IE exclusive setActive() method
could be implemented to override the normal .focus() method
(which attempts to center the viewport around the text field)

Closest solution so far involves saving viewport position prior to giving
focus to new field and returning to it after focus change, but this
results in a jittery behaviour as the browser still attempts the
centering associated with focus.

Will revisit at a later date if a good solution or idea arises
Post asking for assistance:
http://stackoverflow.com/questions/28956790/active-fields-without-scrolling